move template implement into .h file; add range func; add more type convertion support

+++ main.cpp
@@ -3,7 +3,7 @@
 int main()
 {
 	Matrix<float> A = rand(3, 3);
-	Matrix<int> B = {{1,2,3},
+	Matrix<float> B = {{1,2,3},
 					 {3,8,3},
 					 {2,5,6}};
 	cout << A + Matrix<float>(B) << endl;
